Move and export adapter implementations
