+++ packages/sdk/src/core/index.ts
@@ -4,7 +4,9 @@
  */
 
 export * from './context';
+export * from './multipeerAdapter';
 export * from './parameterSet';
 export * from './performanceStats';
 export * from './primitiveTypes';
 export * from './rpc';
+export * from './websocketAdapter';
