--- packages/sdk/src/internal/adapters/websocket/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from './adapter';
